+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# Unreleased
+
+- Add basic example codes.
+- Mark `data_mut()` methods as `unsafe fn` for `GliTexture` and `GliImage`.
+
+
+
 # Version 0.3.1 (2019-4-7)
 
 - Fix build error on Windows platform(Temporarily disable some API on Windows).